fix failing unit test
