+++ src/applications/vaos/tests/services/appointment/index.v2.unit.spec.js
@@ -1,7 +1,6 @@
 /* eslint-disable @department-of-veterans-affairs/axe-check-required */
 import { expect } from 'chai';
 import sinon from 'sinon';
-import { diff } from 'just-diff';
 import {
   mockFetch,
   setFetchJSONFailure,
@@ -9,22 +8,32 @@ import {
 } from '@department-of-veterans-affairs/platform-testing/helpers';
 
 import {
+  FUTURE_APPOINTMENTS_HIDDEN_SET,
   fetchBookedAppointment,
   getAppointmentRequests,
   getLongTermAppointmentHistoryV2,
+  isUpcomingAppointmentOrRequest,
+  isValidPastAppointment,
 } from '../../../services/appointment';
-import { VIDEO_TYPES } from '../../../utils/constants';
+import {
+  APPOINTMENT_STATUS,
+  APPOINTMENT_TYPES,
+  VIDEO_TYPES,
+} from '../../../utils/constants';
 import moment from '../../../lib/moment-tz';
 import { createMockAppointmentByVersion } from '../../mocks/data';
 import {
   getDateRanges,
   mockVAOSAppointmentsFetch,
 } from '../../mocks/helpers.v2';
+import { generateAppointmentUrl } from '../../../utils/appointment';
+import { setRequestedPeriod } from '../../mocks/helpers';
 
 describe('VAOS Appointment service', () => {
-  describe('fetchBookedAppointment', () => {
+  describe('fetchBookedAppointment by id', () => {
     beforeEach(() => mockFetch());
-    it('should return matching v0 and v2 data for an in person VA appointment', async () => {
+
+    it('should return data for an in person VA appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -43,49 +52,16 @@ describe('VAOS Appointment service', () => {
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
         { data: createMockAppointmentByVersion({ version: 2, ...data }) },
       );
-      setFetchJSONResponse(
-        global.fetch.withArgs(
-          sinon.match(`/vaos/v0/appointments/va/${data.id}`),
-        ),
-        { data: createMockAppointmentByVersion({ version: 0, ...data }) },
-      );
 
-      const [v2Result, v0Result] = await Promise.all([
-        fetchBookedAppointment({
-          id: data.id,
-          type: 'va',
-          useV2: true,
-        }),
-        fetchBookedAppointment({ id: data.id, type: 'va' }),
-      ]);
+      const v2Result = await fetchBookedAppointment({
+        id: data.id,
+        useV2: true,
+      });
 
-      // These are always different
-      delete v0Result.vaos.apiData;
-      delete v2Result.vaos.apiData;
-      delete v0Result.version;
-      delete v2Result.version;
-
-      // differences format is http://jsonpatch.com/
-      const differences = diff(v2Result, v0Result);
-      expect(differences).to.have.deep.members(
-        [
-          // The v2 endpoint doesn't send us the clinic name
-          { op: 'remove', path: ['practitioners'] },
-          { op: 'remove', path: ['preferredProviderName'] },
-          { op: 'remove', path: ['vaos', 'facilityData'] },
-          { op: 'replace', path: ['description'], value: 'FUTURE' },
-          { op: 'replace', path: ['comment'], value: null },
-          {
-            op: 'replace',
-            path: ['location', 'clinicName'],
-            value: 'Friendly clinic name',
-          }, // The v2 endpoint doesn't send us the vista status
-        ],
-        'Transformers for v0 and v2 appointment data are out of sync',
-      );
+      expect(v2Result).to.be.ok;
     });
 
-    it('should return matching v0 and v2 data for a cancelled VA appointment', async () => {
+    it('should return data for a cancelled VA appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -102,49 +78,16 @@ describe('VAOS Appointment service', () => {
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
         { data: createMockAppointmentByVersion({ version: 2, ...data }) },
       );
-      setFetchJSONResponse(
-        global.fetch.withArgs(
-          sinon.match(`/vaos/v0/appointments/va/${data.id}`),
-        ),
-        { data: createMockAppointmentByVersion({ version: 0, ...data }) },
-      );
 
-      const [v2Result, v0Result] = await Promise.all([
-        fetchBookedAppointment({
-          id: data.id,
-          type: 'va',
-          useV2: true,
-        }),
-        fetchBookedAppointment({ id: data.id, type: 'va' }),
-      ]);
+      const v2Result = await fetchBookedAppointment({
+        id: data.id,
+        useV2: true,
+      });
 
-      // These are always different
-      delete v0Result.vaos.apiData;
-      delete v2Result.vaos.apiData;
-      delete v0Result.version;
-      delete v2Result.version;
-
-      // differences format is http://jsonpatch.com/
-      const differences = diff(v2Result, v0Result);
-      expect(differences).to.have.deep.members(
-        [
-          // The v2 endpoint doesn't send us the vista status
-          { op: 'remove', path: ['practitioners'] },
-          { op: 'remove', path: ['preferredProviderName'] },
-          { op: 'remove', path: ['vaos', 'facilityData'] },
-          {
-            op: 'replace',
-            path: ['description'],
-            value: 'CANCELLED BY PATIENT',
-          },
-          { op: 'replace', path: ['comment'], value: null },
-          { op: 'replace', path: ['cancelationReason'], value: 'pat' },
-        ],
-        'Transformers for v0 and v2 appointment data are out of sync',
-      );
+      expect(v2Result).to.be.ok;
     });
 
-    it('should return matching v0 and v2 data for a past VA appointment', async () => {
+    it('should return data for a past VA appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -161,44 +104,16 @@ describe('VAOS Appointment service', () => {
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
         { data: createMockAppointmentByVersion({ version: 2, ...data }) },
       );
-      setFetchJSONResponse(
-        global.fetch.withArgs(
-          sinon.match(`/vaos/v0/appointments/va/${data.id}`),
-        ),
-        { data: createMockAppointmentByVersion({ version: 0, ...data }) },
-      );
 
-      const [v2Result, v0Result] = await Promise.all([
-        fetchBookedAppointment({
-          id: data.id,
-          type: 'va',
-          useV2: true,
-        }),
-        fetchBookedAppointment({ id: data.id, type: 'va' }),
-      ]);
+      const v2Result = await fetchBookedAppointment({
+        id: data.id,
+        useV2: true,
+      });
 
-      // These are always different
-      delete v0Result.vaos.apiData;
-      delete v2Result.vaos.apiData;
-      delete v0Result.version;
-      delete v2Result.version;
-
-      // differences format is http://jsonpatch.com/
-      const differences = diff(v2Result, v0Result);
-      expect(differences).to.have.deep.members(
-        [
-          // The v2 endpoint doesn't send us the vista status
-          { op: 'remove', path: ['practitioners'] },
-          { op: 'remove', path: ['preferredProviderName'] },
-          { op: 'remove', path: ['vaos', 'facilityData'] },
-          { op: 'replace', path: ['description'], value: 'CHECKED OUT' },
-          { op: 'replace', path: ['comment'], value: null },
-        ],
-        'Transformers for v0 and v2 appointment data are out of sync',
-      );
+      expect(v2Result).to.be.ok;
     });
 
-    it('should return matching v0 and v2 data for a VA phone appointment', async () => {
+    it('should return data for a VA phone appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -215,44 +130,16 @@ describe('VAOS Appointment service', () => {
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
         { data: createMockAppointmentByVersion({ version: 2, ...data }) },
       );
-      setFetchJSONResponse(
-        global.fetch.withArgs(
-          sinon.match(`/vaos/v0/appointments/va/${data.id}`),
-        ),
-        { data: createMockAppointmentByVersion({ version: 0, ...data }) },
-      );
 
-      const [v2Result, v0Result] = await Promise.all([
-        fetchBookedAppointment({
-          id: data.id,
-          type: 'va',
-          useV2: true,
-        }),
-        fetchBookedAppointment({ id: data.id, type: 'va' }),
-      ]);
+      const v2Result = await fetchBookedAppointment({
+        id: data.id,
+        useV2: true,
+      });
 
-      // These are always different
-      delete v0Result.vaos.apiData;
-      delete v2Result.vaos.apiData;
-      delete v0Result.version;
-      delete v2Result.version;
-
-      // differences format is http://jsonpatch.com/
-      const differences = diff(v2Result, v0Result);
-      expect(differences).to.have.deep.members(
-        [
-          // The v2 endpoint doesn't send us the vista status
-          { op: 'remove', path: ['practitioners'] },
-          { op: 'remove', path: ['preferredProviderName'] },
-          { op: 'remove', path: ['vaos', 'facilityData'] },
-          { op: 'replace', path: ['description'], value: 'CHECKED OUT' },
-          { op: 'replace', path: ['comment'], value: null },
-        ],
-        'Transformers for v0 and v2 appointment data are out of sync',
-      );
+      expect(v2Result).to.be.ok;
     });
 
-    it('should return matching v0 and v2 data for a VA covid vaccine appointment', async () => {
+    it('should return data for a VA covid vaccine appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -270,44 +157,16 @@ describe('VAOS Appointment service', () => {
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
         { data: createMockAppointmentByVersion({ version: 2, ...data }) },
       );
-      setFetchJSONResponse(
-        global.fetch.withArgs(
-          sinon.match(`/vaos/v0/appointments/va/${data.id}`),
-        ),
-        { data: createMockAppointmentByVersion({ version: 0, ...data }) },
-      );
 
-      const [v2Result, v0Result] = await Promise.all([
-        fetchBookedAppointment({
-          id: data.id,
-          type: 'va',
-          useV2: true,
-        }),
-        fetchBookedAppointment({ id: data.id, type: 'va' }),
-      ]);
+      const v2Result = await fetchBookedAppointment({
+        id: data.id,
+        useV2: true,
+      });
 
-      // These are always different
-      delete v0Result.vaos.apiData;
-      delete v2Result.vaos.apiData;
-      delete v0Result.version;
-      delete v2Result.version;
-
-      // differences format is http://jsonpatch.com/
-      const differences = diff(v2Result, v0Result);
-      expect(differences).to.have.deep.members(
-        [
-          // The v2 endpoint doesn't send us the vista status
-          { op: 'remove', path: ['practitioners'] },
-          { op: 'remove', path: ['preferredProviderName'] },
-          { op: 'remove', path: ['vaos', 'facilityData'] },
-          { op: 'replace', path: ['description'], value: 'CHECKED OUT' },
-          { op: 'replace', path: ['comment'], value: null },
-        ],
-        'Transformers for v0 and v2 appointment data are out of sync',
-      );
+      expect(v2Result).to.be.ok;
     });
 
-    it('should return matching v0 and v2 data for a VVC at home video appointment', async () => {
+    it('should return data for a VVC at home video appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -336,57 +195,16 @@ describe('VAOS Appointment service', () => {
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
         { data: createMockAppointmentByVersion({ version: 2, ...data }) },
       );
-      setFetchJSONResponse(
-        global.fetch.withArgs(
-          sinon.match(`/vaos/v0/appointments/va/${data.id}`),
-        ),
-        { data: createMockAppointmentByVersion({ version: 0, ...data }) },
-      );
 
-      const [v2Result, v0Result] = await Promise.all([
-        fetchBookedAppointment({
-          id: data.id,
-          type: 'va',
-          useV2: true,
-        }),
-        fetchBookedAppointment({ id: data.id, type: 'va' }),
-      ]);
-      // These are always different
-      delete v0Result.vaos.apiData;
-      delete v2Result.vaos.apiData;
-      delete v0Result.version;
-      delete v2Result.version;
+      const v2Result = await fetchBookedAppointment({
+        id: data.id,
+        useV2: true,
+      });
 
-      // differences format is http://jsonpatch.com/
-      const differences = diff(v2Result, v0Result);
-      expect(differences).to.have.deep.members(
-        [
-          // The v2 endpoint doesn't send us the vista status
-          { op: 'remove', path: ['practitioners'] },
-          { op: 'remove', path: ['preferredProviderName'] },
-          { op: 'remove', path: ['vaos', 'facilityData'] },
-          { op: 'replace', path: ['description'], value: 'FUTURE' },
-          {
-            op: 'replace',
-            path: ['videoData', 'providers', 0, 'name', 'firstName'],
-            value: undefined,
-          },
-          {
-            op: 'replace',
-            path: ['videoData', 'providers', 0, 'name', 'lastName'],
-            value: undefined,
-          },
-          {
-            op: 'replace',
-            path: ['videoData', 'providers', 0, 'display'],
-            value: 'undefined undefined',
-          },
-        ],
-        'Transformers for v0 and v2 appointment data are out of sync',
-      );
+      expect(v2Result).to.be.ok;
     });
 
-    it('should return matching v0 and v2 data for an ATLAS video appointment', async () => {
+    it('should return data for an ATLAS video appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -428,57 +246,16 @@ describe('VAOS Appointment service', () => {
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
         { data: createMockAppointmentByVersion({ version: 2, ...data }) },
       );
-      setFetchJSONResponse(
-        global.fetch.withArgs(
-          sinon.match(`/vaos/v0/appointments/va/${data.id}`),
-        ),
-        { data: createMockAppointmentByVersion({ version: 0, ...data }) },
-      );
 
-      const [v2Result, v0Result] = await Promise.all([
-        fetchBookedAppointment({
-          id: data.id,
-          type: 'va',
-          useV2: true,
-        }),
-        fetchBookedAppointment({ id: data.id, type: 'va' }),
-      ]);
-      // These are always different
-      delete v0Result.vaos.apiData;
-      delete v2Result.vaos.apiData;
-      delete v0Result.version;
-      delete v2Result.version;
+      const v2Result = await fetchBookedAppointment({
+        id: data.id,
+        useV2: true,
+      });
 
-      // differences format is http://jsonpatch.com/
-      const differences = diff(v2Result, v0Result);
-      expect(differences).to.have.deep.members(
-        [
-          // The v2 endpoint doesn't send us the vista status
-          { op: 'remove', path: ['practitioners'] },
-          { op: 'remove', path: ['preferredProviderName'] },
-          { op: 'remove', path: ['vaos', 'facilityData'] },
-          { op: 'replace', path: ['description'], value: 'FUTURE' },
-          {
-            op: 'replace',
-            path: ['videoData', 'providers', 0, 'name', 'firstName'],
-            value: undefined,
-          },
-          {
-            op: 'replace',
-            path: ['videoData', 'providers', 0, 'name', 'lastName'],
-            value: undefined,
-          },
-          {
-            op: 'replace',
-            path: ['videoData', 'providers', 0, 'display'],
-            value: 'undefined undefined',
-          },
-        ],
-        'Transformers for v0 and v2 appointment data are out of sync',
-      );
+      expect(v2Result).to.be.ok;
     });
 
-    it('should return matching v0 and v2 data for a clinic based video appointment', async () => {
+    it('should return data for a clinic based video appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -499,42 +276,16 @@ describe('VAOS Appointment service', () => {
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
         { data: createMockAppointmentByVersion({ version: 2, ...data }) },
       );
-      setFetchJSONResponse(
-        global.fetch.withArgs(
-          sinon.match(`/vaos/v0/appointments/va/${data.id}`),
-        ),
-        { data: createMockAppointmentByVersion({ version: 0, ...data }) },
-      );
 
-      const [v2Result, v0Result] = await Promise.all([
-        fetchBookedAppointment({
-          id: data.id,
-          type: 'va',
-          useV2: true,
-        }),
-        fetchBookedAppointment({ id: data.id, type: 'va' }),
-      ]);
-      // These are always different
-      delete v0Result.vaos.apiData;
-      delete v2Result.vaos.apiData;
-      delete v0Result.version;
-      delete v2Result.version;
+      const v2Result = await fetchBookedAppointment({
+        id: data.id,
+        useV2: true,
+      });
 
-      // differences format is http://jsonpatch.com/
-      const differences = diff(v2Result, v0Result);
-      expect(differences).to.have.deep.members(
-        [
-          // The v2 endpoint doesn't send us the vista status
-          { op: 'replace', path: ['description'], value: 'FUTURE' },
-          { op: 'remove', path: ['practitioners'] },
-          { op: 'remove', path: ['preferredProviderName'] },
-          { op: 'remove', path: ['vaos', 'facilityData'] },
-        ],
-        'Transformers for v0 and v2 appointment data are out of sync',
-      );
+      expect(v2Result).to.be.ok;
     });
 
-    it('should return matching v0 and v2 data for a store forward video appointment', async () => {
+    it('should return data for a store forward video appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -555,42 +306,16 @@ describe('VAOS Appointment service', () => {
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
         { data: createMockAppointmentByVersion({ version: 2, ...data }) },
       );
-      setFetchJSONResponse(
-        global.fetch.withArgs(
-          sinon.match(`/vaos/v0/appointments/va/${data.id}`),
-        ),
-        { data: createMockAppointmentByVersion({ version: 0, ...data }) },
-      );
 
-      const [v2Result, v0Result] = await Promise.all([
-        fetchBookedAppointment({
-          id: data.id,
-          type: 'va',
-          useV2: true,
-        }),
-        fetchBookedAppointment({ id: data.id, type: 'va' }),
-      ]);
-      // These are always different
-      delete v0Result.vaos.apiData;
-      delete v2Result.vaos.apiData;
-      delete v0Result.version;
-      delete v2Result.version;
+      const v2Result = await fetchBookedAppointment({
+        id: data.id,
+        useV2: true,
+      });
 
-      // differences format is http://jsonpatch.com/
-      const differences = diff(v2Result, v0Result);
-      expect(differences).to.have.deep.members(
-        [
-          // The v2 endpoint doesn't send us the vista status
-          { op: 'replace', path: ['description'], value: 'FUTURE' },
-          { op: 'remove', path: ['practitioners'] },
-          { op: 'remove', path: ['preferredProviderName'] },
-          { op: 'remove', path: ['vaos', 'facilityData'] },
-        ],
-        'Transformers for v0 and v2 appointment data are out of sync',
-      );
+      expect(v2Result).to.be.ok;
     });
 
-    it('should return matching v0 and v2 data for a mobile any video appointment', async () => {
+    it('should return data for a mobile any video appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -609,43 +334,16 @@ describe('VAOS Appointment service', () => {
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
         { data: createMockAppointmentByVersion({ version: 2, ...data }) },
       );
-      setFetchJSONResponse(
-        global.fetch.withArgs(
-          sinon.match(`/vaos/v0/appointments/va/${data.id}`),
-        ),
-        { data: createMockAppointmentByVersion({ version: 0, ...data }) },
-      );
 
-      const [v2Result, v0Result] = await Promise.all([
-        fetchBookedAppointment({
-          id: data.id,
-          type: 'va',
-          useV2: true,
-        }),
-        fetchBookedAppointment({ id: data.id, type: 'va' }),
-      ]);
+      const v2Result = await fetchBookedAppointment({
+        id: data.id,
+        useV2: true,
+      });
 
-      // These are always different
-      delete v0Result.vaos.apiData;
-      delete v2Result.vaos.apiData;
-      delete v0Result.version;
-      delete v2Result.version;
-
-      // differences format is http://jsonpatch.com/
-      const differences = diff(v2Result, v0Result);
-      expect(differences).to.have.deep.members(
-        [
-          // The v2 endpoint doesn't send us the vista status
-          { op: 'replace', path: ['description'], value: 'FUTURE' },
-          { op: 'remove', path: ['practitioners'] },
-          { op: 'remove', path: ['preferredProviderName'] },
-          { op: 'remove', path: ['vaos', 'facilityData'] },
-        ],
-        'Transformers for v0 and v2 appointment data are out of sync',
-      );
+      expect(v2Result).to.be.ok;
     });
 
-    it('should return matching v0 and v2 data for a mobile gfe video appointment', async () => {
+    it('should return data for a mobile gfe video appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -664,43 +362,17 @@ describe('VAOS Appointment service', () => {
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
         { data: createMockAppointmentByVersion({ version: 2, ...data }) },
       );
-      setFetchJSONResponse(
-        global.fetch.withArgs(
-          sinon.match(`/vaos/v0/appointments/va/${data.id}`),
-        ),
-        { data: createMockAppointmentByVersion({ version: 0, ...data }) },
-      );
 
-      const [v2Result, v0Result] = await Promise.all([
-        fetchBookedAppointment({
-          id: data.id,
-          type: 'va',
-          useV2: true,
-        }),
-        fetchBookedAppointment({ id: data.id, type: 'va' }),
-      ]);
+      const v2Result = await fetchBookedAppointment({
+        id: data.id,
 
-      // These are always different
-      delete v0Result.vaos.apiData;
-      delete v2Result.vaos.apiData;
-      delete v0Result.version;
-      delete v2Result.version;
+        useV2: true,
+      });
 
-      // differences format is http://jsonpatch.com/
-      const differences = diff(v2Result, v0Result);
-      expect(differences).to.have.deep.members(
-        [
-          // The v2 endpoint doesn't send us the vista status
-          { op: 'replace', path: ['description'], value: 'FUTURE' },
-          { op: 'remove', path: ['practitioners'] },
-          { op: 'remove', path: ['preferredProviderName'] },
-          { op: 'remove', path: ['vaos', 'facilityData'] },
-        ],
-        'Transformers for v0 and v2 appointment data are out of sync',
-      );
+      expect(v2Result).to.be.ok;
     });
 
-    it('should return matching v0 and v2 data for a community care appointment', async () => {
+    it('should return data for a community care appointment', async () => {
       const data = {
         id: '1234',
         start: moment()
@@ -730,80 +402,16 @@ describe('VAOS Appointment service', () => {
         global.fetch.withArgs(sinon.match(`/vaos/v2/appointments/${data.id}`)),
         { data: createMockAppointmentByVersion({ version: 2, ...data }) },
       );
-      setFetchJSONResponse(
-        global.fetch.withArgs(sinon.match(`/vaos/v0/appointments`)),
-        { data: [createMockAppointmentByVersion({ version: 0, ...data })] },
-      );
 
-      const [v2Result, v0Result] = await Promise.all([
-        fetchBookedAppointment({
-          id: data.id,
-          type: 'cc',
-          useV2: true,
-        }),
-        fetchBookedAppointment({ id: data.id, type: 'cc' }),
-      ]);
-      // These are always different
-      delete v0Result.vaos.apiData;
-      delete v2Result.vaos.apiData;
-      delete v0Result.version;
-      delete v2Result.version;
+      const v2Result = await fetchBookedAppointment({
+        id: data.id,
+        useV2: true,
+      });
 
-      // The CC date transformer logic sets the date in UTC mode, which creates
-      // a format difference when this test is run on a machine in GMT/UTC
-      // This adjusts for that format difference, because fixing the code results
-      // in weird test failures that need to be looked at separately
-      v0Result.start = v0Result.start.replace('Z', '+00:00');
-
-      // differences format is http://jsonpatch.com/
-      const differences = diff(v2Result, v0Result);
-      expect(differences).to.have.deep.members(
-        [
-          {
-            op: 'remove',
-            path: ['communityCareProvider', 'providers'],
-          },
-          {
-            op: 'remove',
-            path: ['communityCareProvider', 'treatmentSpecialty'],
-          },
-          { op: 'remove', path: ['practitioners'] },
-          { op: 'remove', path: ['preferredProviderName'] },
-          { op: 'remove', path: ['vaos', 'facilityData'] },
-          { op: 'replace', path: ['comment'], value: null },
-          {
-            op: 'replace',
-            path: ['communityCareProvider', 'telecom'],
-            value: null,
-          },
-          {
-            op: 'replace',
-            path: ['vaos', 'timeZone'],
-            value: moment(data.start)
-              .tz(moment.tz.guess())
-              .format('Z z'),
-          },
-          {
-            op: 'add',
-            path: ['communityCareProvider', 'firstName'],
-            value: null,
-          },
-          {
-            op: 'add',
-            path: ['communityCareProvider', 'lastName'],
-            value: null,
-          },
-          {
-            op: 'add',
-            path: ['preferredCommunityCareProviders'],
-            value: null,
-          },
-        ],
-        'Transformers for v0 and v2 appointment data are out of sync',
-      );
+      expect(v2Result).to.be.ok;
     });
 
-    it('should return matching v0 and v2 data for an error fetching appointments', async () => {
+    it('should return data for an error fetching appointments', async () => {
       const error = {
         code: 'VAOS_504',
         title: 'Gateway error',
@@ -816,42 +424,25 @@ describe('VAOS Appointment service', () => {
           errors: [error],
         },
       );
-      setFetchJSONFailure(
-        global.fetch.withArgs(sinon.match(`/vaos/v0/appointments/va/1234`)),
-        {
-          errors: [error],
-        },
-      );
 
-      let v0Result = null;
       let v2Result = null;
       try {
         await fetchBookedAppointment({
           id: '1234',
-          type: 'va',
           useV2: true,
         });
-      } catch (e) {
-        v0Result = e;
-      }
-
-      try {
-        await fetchBookedAppointment({ id: '1234', type: 'va' });
       } catch (e) {
         v2Result = e;
       }
 
       expect(v2Result).to.be.ok;
-      expect(v0Result).to.be.ok;
-
-      const differences = diff(v2Result, v0Result);
-      expect(differences).to.be.empty;
     });
   });
 
   describe('getAppointmentRequests', () => {
     beforeEach(() => mockFetch());
-    it('should return matching v0 and v2 data for a VA appointment request', async () => {
+
+    it('should return data for a VA appointment request', async () => {
       // Given VA appointment request
       const data = {
         id: '1234',
@@ -875,87 +466,43 @@ describe('VAOS Appointment service', () => {
       };
       const startDate = moment()
         .subtract(30, 'days')
-        .format();
+        .format('YYYY-MM-DD');
       const endDate = moment()
         .add(30, 'days')
-        .format();
+        .format('YYYY-MM-DD');
 
-      // And the developer fetched that request through both the v2 and v0 APIs
-      setFetchJSONResponse(
-        global.fetch.withArgs(
-          sinon.match(
-            `/vaos/v2/appointments?_include=facilities,clinics&start=${startDate}&end=${endDate}&statuses[]=proposed&statuses[]=cancelled`,
-          ),
-        ),
-        {
-          data: [
-            createMockAppointmentByVersion({
-              version: 2,
-              ...data,
-              requestedPeriods: [
-                {
-                  start: `${moment().format('YYYY-MM-DD')}T00:00:00Z`,
-                  end: `${moment().format('YYYY-MM-DD')}T11:59:59Z`,
-                },
-              ],
-            }),
-          ],
-        },
-      );
-
-      setFetchJSONResponse(
-        global.fetch.withArgs(
-          sinon.match(
-            `/vaos/v0/appointment_requests?start_date=${startDate}&end_date=${endDate}`,
-          ),
-        ),
-        {
-          data: [
-            createMockAppointmentByVersion({
-              version: 0,
-              ...data,
-            }),
-          ],
-        },
-      );
-
-      const [v0Result, v2Result] = await Promise.all([
-        getAppointmentRequests({
-          startDate,
-          endDate,
-        }),
-        getAppointmentRequests({
-          startDate,
-          endDate,
-          useV2: true,
-        }),
+      // And the developer fetched that request through both the v2 APIs
+      const url = generateAppointmentUrl(startDate, endDate, [
+        'proposed',
+        'cancelled',
       ]);
 
-      // These are always different
-      delete v0Result[0].vaos.apiData;
-      delete v2Result[0].vaos.apiData;
-      delete v0Result[0].version;
-      delete v2Result[0].version;
-
-      // When they compare the two results
-      // differences format is http://jsonpatch.com/
-      const differences = diff(v2Result[0], v0Result[0]);
-
-      // Then the results have the following differences
-      expect(differences).to.have.deep.members(
-        [
-          { op: 'remove', path: ['description'] },
-          { op: 'remove', path: ['practitioners'] },
-          { op: 'remove', path: ['preferredProviderName'] },
-          { op: 'remove', path: ['vaos', 'facilityData'] },
-          { op: 'replace', path: ['reason'], value: undefined },
-          { op: 'replace', path: ['comment'], value: null },
+      setFetchJSONResponse(global.fetch.withArgs(sinon.match(url)), {
+        data: [
+          createMockAppointmentByVersion({
+            version: 2,
+            ...data,
+            requestedPeriods: [
+              {
+                start: `${moment().format('YYYY-MM-DD')}T00:00:00Z`,
+                end: `${moment().format('YYYY-MM-DD')}T11:59:59Z`,
+              },
+            ],
+          }),
         ],
-        'Transformers for v0 and v2 appointment request data are out of sync',
-      );
+      });
+
+      const v2Result = await getAppointmentRequests({
+        startDate,
+        endDate,
+        useV2: true,
+      });
+
+      // Then expect a VA appointment request to be returned.
+      expect(v2Result.length).to.be.gt(0);
     });
 
-    it('should return matching v0 and v2 data for a CC appointment request', async () => {
+    it('should return data for a CC appointment request', async () => {
       // Given CC appointment request
       const data = {
         id: '1234',
@@ -1004,70 +551,17 @@ describe('VAOS Appointment service', () => {
         },
       );
 
-      setFetchJSONResponse(
-        global.fetch.withArgs(
-          sinon.match(
-            `/vaos/v0/appointment_requests?start_date=${startDate}&end_date=${endDate}`,
-          ),
-        ),
-        {
-          data: [
-            createMockAppointmentByVersion({
-              version: 0,
-              ...data,
-            }),
-          ],
-        },
-      );
+      const v2Result = await getAppointmentRequests({
+        startDate,
+        endDate,
+        useV2: true,
+      });
 
-      const [v0Result, v2Result] = await Promise.all([
-        getAppointmentRequests({
-          startDate,
-          endDate,
-        }),
-        getAppointmentRequests({
-          startDate,
-          endDate,
-          useV2: true,
-        }),
-      ]);
-
-      // These are always different
-      delete v0Result[0].vaos.apiData;
-      delete v2Result[0].vaos.apiData;
-      delete v0Result[0].version;
-      delete v2Result[0].version;
-
-      // When they compare the two results
-      // differences format is http://jsonpatch.com/
-      const differences = diff(v2Result[0], v0Result[0]);
-      // Then the results have the following differences
-      expect(differences).to.have.deep.members(
-        [
-          { op: 'remove', path: ['practitioners'] },
-          { op: 'remove', path: ['preferredProviderName'] },
-          { op: 'remove', path: ['description'] },
-          { op: 'remove', path: ['vaos', 'facilityData'] },
-          { op: 'replace', path: ['reason'], value: undefined },
-          { op: 'replace', path: ['comment'], value: null },
-
-          {
-            op: 'replace',
-            path: ['type', 'coding', 0, 'code'],
-            value: 'CCPRMYRTNE',
-          },
-          { op: 'replace', path: ['requestVisitType'], value: 'Office visit' },
-          {
-            op: 'add',
-            path: ['preferredCommunityCareProviders'],
-            value: null,
-          },
-        ],
-        'Transformers for v0 and v2 appointment request data are out of sync',
-      );
+      // Then expect a CC appointment request to be returned.
+      expect(v2Result.length).to.be.gt(0);
     });
 
-    it('should return matching v0 and v2 data for a cancelled appointment request', async () => {
+    it('should return data for a cancelled appointment request', async () => {
       // Given cancelled VA appointment request
       const data = {
         id: '1234',
@@ -1113,60 +607,17 @@ describe('VAOS Appointment service', () => {
         },
       );
 
-      setFetchJSONResponse(
-        global.fetch.withArgs(
-          sinon.match(
-            `/vaos/v0/appointment_requests?start_date=${startDate}&end_date=${endDate}`,
-          ),
-        ),
-        {
-          data: [
-            createMockAppointmentByVersion({
-              version: 0,
-              ...data,
-            }),
-          ],
-        },
-      );
+      const v2Result = await getAppointmentRequests({
+        startDate,
+        endDate,
+        useV2: true,
+      });
 
-      const [v0Result, v2Result] = await Promise.all([
-        getAppointmentRequests({
-          startDate,
-          endDate,
-        }),
-        getAppointmentRequests({
-          startDate,
-          endDate,
-          useV2: true,
-        }),
-      ]);
-
-      // These are always different
-      delete v0Result[0].vaos.apiData;
-      delete v2Result[0].vaos.apiData;
-      delete v0Result[0].version;
-      delete v2Result[0].version;
-
-      // When they compare the two results
-      // differences format is http://jsonpatch.com/
-      const differences = diff(v2Result[0], v0Result[0]);
-
-      // Then the results have the following differences
-      expect(differences).to.have.deep.members(
-        [
-          // { op: 'remove', path: ['reason'] },
-          { op: 'remove', path: ['practitioners'] },
-          { op: 'remove', path: ['preferredProviderName'] },
-          { op: 'remove', path: ['description'] },
-          { op: 'remove', path: ['vaos', 'facilityData'] },
-          { op: 'replace', path: ['reason'], value: undefined },
-          { op: 'replace', path: ['comment'], value: null },
-        ],
-        'Transformers for v0 and v2 appointment request data are out of sync',
-      );
+      // Then expect a canceled appointment request to be returned.
+      expect(v2Result.length).to.be.gt(0);
     });
 
-    it('should return matching v0 and v2 data for an error fetching requests', async () => {
+    it('should return data for an error fetching requests', async () => {
       // Given VAOS error
       const error = {
         code: 'VAOS_504',
@@ -1194,46 +645,20 @@ describe('VAOS Appointment service', () => {
         },
       );
 
-      setFetchJSONFailure(
-        global.fetch.withArgs(
-          sinon.match(
-            `/vaos/v0/appointment_requests?start_date=${startDate}&end_date=${endDate}`,
-          ),
-        ),
-        {
-          errors: [error],
-        },
-      );
-
-      let v0Result = null;
       let v2Result = null;
 
       try {
         await getAppointmentRequests({
           startDate,
           endDate,
-        });
-      } catch (e) {
-        v0Result = e;
-      }
-
-      try {
-        await getAppointmentRequests({
-          startDate,
-          endDate,
+          useV2: true,
         });
       } catch (e) {
         v2Result = e;
       }
 
+      // Then expect an error to be returned.
       expect(v2Result).to.be.ok;
-      expect(v0Result).to.be.ok;
-
-      // When they compare the two results
-      const differences = diff(v2Result, v0Result);
-
-      // Then the results have the following differences
-      expect(differences).to.be.empty;
     });
   });
 
@@ -1261,6 +686,233 @@ describe('VAOS Appointment service', () => {
       expect(global.fetch.secondCall.args[0]).to.contain(dateRanges[1].end);
       expect(global.fetch.thirdCall.args[0]).to.contain(dateRanges[2].start);
       expect(global.fetch.thirdCall.args[0]).to.contain(dateRanges[2].end);
+    });
+  });
+
+  //--------
+
+  const now = moment();
+  describe('isUpcomingAppointmentOrRequest', () => {
+    it('should filter future requests', () => {
+      const apptRequests = [
+        // canceled past - should filter out
+        {
+          status: APPOINTMENT_STATUS.cancelled,
+          requestedPeriod: [
+            setRequestedPeriod(now.clone().add(-2, 'days'), 'AM'),
+          ],
+          vaos: {
+            isExpressCare: false,
+            appointmentType: APPOINTMENT_TYPES.request,
+          },
+        },
+        // cancelled past - should filter out
+        {
+          status: APPOINTMENT_STATUS.cancelled,
+          requestedPeriod: [
+            setRequestedPeriod(now.clone().subtract(22, 'days'), 'AM'),
+          ],
+          vaos: {
+            isExpressCare: false,
+            appointmentType: APPOINTMENT_TYPES.request,
+          },
+        },
+        // pending past - should filter out
+        {
+          status: APPOINTMENT_STATUS.pending,
+          requestedPeriod: [
+            setRequestedPeriod(now.clone().add(-2, 'days'), 'AM'),
+          ],
+          vaos: {
+            isExpressCare: false,
+            appointmentType: APPOINTMENT_TYPES.request,
+          },
+        },
+        // future within 13 - should not filter out
+        {
+          status: APPOINTMENT_STATUS.pending,
+          requestedPeriod: [
+            setRequestedPeriod(
+              now
+                .clone()
+                .add(395, 'days')
+                .add(-1, 'days'),
+              'AM',
+            ),
+          ],
+          vaos: {
+            isExpressCare: false,
+            appointmentType: APPOINTMENT_TYPES.request,
+          },
+        },
+        // future - should not filter out
+        {
+          status: APPOINTMENT_STATUS.pending,
+          requestedPeriod: [
+            setRequestedPeriod(now.clone().add(2, 'days'), 'AM'),
+          ],
+          vaos: {
+            isExpressCare: false,
+            appointmentType: APPOINTMENT_TYPES.request,
+          },
+        },
+        // future canceled - should not filter out
+        {
+          status: APPOINTMENT_STATUS.cancelled,
+          requestedPeriod: [
+            setRequestedPeriod(now.clone().add(3, 'days'), 'AM'),
+          ],
+          vaos: {
+            isExpressCare: false,
+            appointmentType: APPOINTMENT_TYPES.request,
+          },
+        },
+      ];
+
+      const filteredRequests = apptRequests.filter(
+        isUpcomingAppointmentOrRequest,
+      );
+
+      expect(
+        filteredRequests.filter(
+          req => req.status === APPOINTMENT_STATUS.cancelled,
+        ).length,
+      ).to.equal(1);
+      expect(
+        filteredRequests.filter(
+          req => req.status === APPOINTMENT_STATUS.pending,
+        ).length,
+      ).to.equal(3);
+      expect(
+        filteredRequests.filter(req => req.status === 'Booked').length,
+      ).to.equal(0);
+    });
+
+    it('should filter future confirmed appointments', () => {
+      const confirmedAppts = [
+        // appointment more than 395 days should not show
+        {
+          start: '2099-04-30T05:35:00',
+          facilityId: '984',
+          vaos: { appointmentType: APPOINTMENT_TYPES.vaAppointment },
+        },
+        // appointment less than 395 days should show
+        {
+          start: now
+            .clone()
+            .add(394, 'days')
+            .format(),
+          facilityId: '984',
+          vaos: { appointmentType: APPOINTMENT_TYPES.vaAppointment },
+        },
+        // appointment 30 min ago should show
+        {
+          start: now
+            .clone()
+            .subtract(30, 'minutes')
+            .format(),
+          facilityId: '984',
+          vaos: { appointmentType: APPOINTMENT_TYPES.vaAppointment },
+        },
+        // appointment more than 1 hour ago should not show
+        {
+          start: now
+            .clone()
+            .subtract(65, 'minutes')
+            .format(),
+          facilityId: '984',
+          vaos: {
+            isPastAppointment: true,
+            appointmentType: APPOINTMENT_TYPES.vaAppointment,
+          },
+        },
+        // video appointment less than 4 hours ago should show
+        {
+          start: now
+            .clone()
+            .subtract(230, 'minutes')
+            .format(),
+          vaos: {
+            appointmentType: APPOINTMENT_TYPES.vaAppointment,
+          },
+        },
+        // video appointment more than 4 hours ago should not show
+        {
+          start: now
+            .clone()
+            .subtract(245, 'minutes')
+            .format(),
+          vaos: {
+            isPastAppointment: true,
+            appointmentType: APPOINTMENT_TYPES.vaAppointment,
+          },
+        },
+        // appointment with status 'NO-SHOW' should not show
+        {
+          description: 'NO-SHOW',
+          vaos: { appointmentType: APPOINTMENT_TYPES.vaAppointment },
+        },
+        // appointment with status 'DELETED' should not show
+        {
+          description: 'DELETED',
+          vaos: { appointmentType: APPOINTMENT_TYPES.vaAppointment },
+        },
+      ];
+
+      const filteredConfirmed = confirmedAppts.filter(
+        isUpcomingAppointmentOrRequest,
+      );
+      expect(filteredConfirmed.length).to.equal(3);
+    });
+
+    it('should filter out appointments with status in FUTURE_APPOINTMENTS_HIDDEN_SET', () => {
+      const hiddenAppts = [...FUTURE_APPOINTMENTS_HIDDEN_SET].map(
+        currentStatus => ({
+          description: currentStatus,
+          vaos: { appointmentType: APPOINTMENT_TYPES.vaAppointment },
+          facilityId: '984',
+        }),
+      );
+
+      const filtered = hiddenAppts.filter(isUpcomingAppointmentOrRequest);
+      expect(hiddenAppts.length).to.equal(2);
+      expect(filtered.length).to.equal(0);
+    });
+
+    it('should filter out video appointments with status in FUTURE_APPOINTMENTS_HIDDEN_SET', () => {
+      const hiddenAppts = [...FUTURE_APPOINTMENTS_HIDDEN_SET].map(code => ({
+        description: code,
+        vaos: {
+          appointmentType: APPOINTMENT_TYPES.vaAppointment,
+        },
+      }));
+
+      const filtered = hiddenAppts.filter(isUpcomingAppointmentOrRequest);
+      expect(hiddenAppts.length).to.equal(2);
+      expect(filtered.length).to.equal(0);
+    });
+  });
+
+  // NOTE: Refactored v0 test
+  describe('filterPastAppointments', () => {
+    it('should not filter appointments that are not in hidden status set', () => {
+      const appointments = [
+        {
+          vaos: { appointmentType: APPOINTMENT_TYPES.vaAppointment },
+        },
+        {
+          description: 'NO-SHOW',
+          vaos: { appointmentType: APPOINTMENT_TYPES.vaAppointment },
+        },
+        {
+          description: 'CHECKED IN',
+          vaos: { appointmentType: APPOINTMENT_TYPES.vaAppointment },
+        },
+      ];
+
+      const filtered = appointments.filter(isValidPastAppointment);
+
+      expect(filtered.length).to.equal(3);
     });
   });
 });
